+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## Version 0.9.6
+Tested with engines 6.0.15 and above
+
+### Added:
+- dxmc job report --error_details to print detailed information about job failurs or warnings
+
+### Changed:
+- masking jobs with non-conforming data are now displayed with WARNING status
+- Fix for advenced JDBC config for MS SQL connector
+- Various bugs
+
 ## Version 0.9.5
 Tested with engines 6.0.12 and above